% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biogeonet.R
\docType{package}
\name{biogeonet}
\alias{biogeonet}
\title{biogeonet}
\description{
Research compendium of 'On the integration of biotic interaction and environmental constraints at the biogeographical scale' (DOI:10.1111/ecog.01714).
}
\author{
Kevin Cazelles \href{mailto:kevin.cazelles@gmail.com}{kevin.cazelles@gmail.com}
}
