% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getEigenElements}
\alias{getEigenElements}
\title{Markov transition matrix solver.}
\arguments{
\item{markov}{A transition matrix of a markov chain.}
}
\value{
Returns the eigen values associated to the vector and eigen values
transition matrix of the markov chain to be solved.
}
\description{
Solve the Markov transition matrix and return the eigen vector giving
the probability of presence associated with all the networks's species.
}
\author{
Kevin Cazelles
}
