% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nicheModel}
\alias{nicheModel}
\title{Generate an interaction network using the niche model.}
\usage{
nicheModel(nbsp, connec, mode)
}
\arguments{
\item{nbsp}{an integer, the number of species.}

\item{connec}{a real number giving the connectance of the network to be generated (between 0 and .5).}

\item{mode}{an integer used to select the sign used in the network (1- for predation; 2-competition;  3-nutualism)  a real number giving the optimum value, i.e the abiotic values at which the colonisation rate is maximal.}
}
\value{
Returns a numeric matrix describing the interaction network.
}
\description{
This function returns an interaction network using the model
proposed by Williams & Martinez (2000), the sign of coefficients can be
modified to simulate different kind of network.
}
\references{
Williams, R. J. & Martinez, N. D. Simple rules yield complex food webs.
Nature 404, 180-183 (2000).
}
\author{
Kevin Cazelles
}
