% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{solveMarkov}
\alias{solveMarkov}
\title{Markov transition matrix solver.
the probability of presence associated with all species in the network.}
\description{
Solve the Markov transition matrix and return the eigen vector giving
the probability of presence associated with all the networks's species.
}
\author{
Kevin Cazelles
}
