% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{buildMarkov}
\alias{buildMarkov}
\title{Generates the transiton matrix for a given probability and  an interaction network using the niche model.}
\usage{
buildMarkov(nbsp, colo, metaweb, basexti, mn, mx, shape)
}
\arguments{
\item{nbsp}{an integer giving the number of species.}

\item{colo}{a vector giving the colonization rates of all species.}

\item{metaweb}{a matrix describing the interaction network.}

\item{basexti}{extinction rate without interaction.}

\item{mn}{a real number giving the minimum rate of extnction.}

\item{mx}{a real number giving the maximum rate of extnction.}

\item{shape}{a positive real number controling the impact of biotic interactions: the larger the value the stronger the impact of interactions.}
}
\value{
Returns the transition matrix of the markov chain to be solved.
}
\description{
Generates the transiton matrix for a given probability and  an interaction network using the niche model.
}
\references{
Cazelles, K., Mouquet, N., Mouillot, D. & Gravel, D. On the integration of
biotic interaction and environmental constraints at the biogeographical scale.
}
\author{
Kevin Cazelles
}
