% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solveMarkov.R
\name{solveMarkov}
\alias{solveMarkov}
\title{solveMarkov}
\usage{
solveMarkov(markov, sparse = TRUE, continuous = FALSE)
}
\arguments{
\item{markov}{a transition matrix of a Markov chain.}

\item{sparse}{a logical. Should a sparse matrix be used? Default value is \code{TRUE} (See details).}

\item{continuous}{a logical. Is \code{markov} a continuous time Markov chain?
Default is set to \code{FALSE} (menaing it is a discrete one).}
}
\value{
A vector of probabilities
}
\description{
Compute the communities probabilities associated at equilibrium for a given
transition matrix.
}
\details{
If \code{sparse} is \code{TRUE}, then only the larger eigen values is
returned (function \code{eigs_gen()} in the armadillo C++ library requires
the number of eigen values to be computed and is currently set to 1 in
\code{getEigenElementsSp()}).
}
